Update notification feature * Split functionality into separate files * Prevent notifications on certain screens
