+++ src/hooks/useNotification.ts
@@ -1,27 +1,13 @@
+/* istanbul ignore file */
+// This file mainly contains code for IO, and unable to be tested in unit tests.
+import { foregroundMessageHandler } from "@/services/notification";
 import notifee, { AndroidImportance } from "@notifee/react-native";
-import {
-  FirebaseAuthTypes,
-  getAuth,
-  onAuthStateChanged,
-} from "@react-native-firebase/auth";
-import {
-  deleteDoc,
-  doc,
-  getFirestore,
-  serverTimestamp,
-  setDoc,
-} from "@react-native-firebase/firestore";
-import {
-  FirebaseMessagingTypes,
-  getMessaging,
-  getToken,
-  onMessage,
-  onTokenRefresh,
-  setBackgroundMessageHandler,
-} from "@react-native-firebase/messaging";
-import { useEffect, useState } from "react";
+import { usePathname } from "expo-router";
+import { useEffect } from "react";
 
 export default function useNotification() {
+  const path = usePathname();
+
   useEffect(() => {
     notifee.requestPermission();
   }, []);
@@ -38,73 +24,6 @@ export default function useNotification() {
   }, []);
 
   useEffect(() => {
-    return foregroundMessageHandler();
-  }, []);
-
-  const [user, setUser] = useState<FirebaseAuthTypes.User | null>(null);
-
-  useEffect(() => {
-    const subscriber = onAuthStateChanged(getAuth(), (user) => {
-      setUser(user);
-    });
-    return subscriber; // unsubscribe on unmount
-  }, []);
-
-  useEffect(() => {
-    if (!user) {
-      return;
-    }
-
-    const uid = user.uid;
-    const m = getMessaging();
-    getToken(m).then((token) => {
-      console.log("FCM Token:", token);
-      registerToken(uid, token);
-    });
-
-    return onTokenRefresh(m, (token) => {
-      registerToken(uid, token);
-    });
-  }, [user]);
-}
-
-export function backgroundMessageHandler() {
-  notifee.onBackgroundEvent(async () => {});
-
-  setBackgroundMessageHandler(getMessaging(), onMessageReceived);
-}
-
-function foregroundMessageHandler(): () => void {
-  return onMessage(getMessaging(), onMessageReceived);
-}
-
-async function onMessageReceived(
-  message: FirebaseMessagingTypes.RemoteMessage,
-) {
-  console.log("FCM Message:", message);
-  const m = message.data?.["notifee"];
-  if (typeof m === "string") {
-    await notifee.displayNotification(JSON.parse(m));
-  }
-}
-
-async function registerToken(uid: string, token: string) {
-  const db = getFirestore();
-  const docRef = doc(db, "notifications", token);
-  console.log("Register token:", token, "for user:", uid);
-  await setDoc(docRef, {
-    uid: uid,
-    token: token,
-    timestamp: serverTimestamp(),
-  });
-}
-
-export async function deregisterToken() {
-  const uid = getAuth().currentUser!.uid;
-  const token = await getToken(getMessaging());
-
-  const db = getFirestore();
-  const docRef = doc(db, "notifications", token);
-  console.log("Deregister token:", token, "for user:", uid);
-  await deleteDoc(docRef);
+    return foregroundMessageHandler(path);
+  }, [path]);
 }
